move t file
