+++ packages/react-core/src/translation/T.tsx
@@ -1,15 +1,15 @@
 import React, { Suspense } from 'react';
-import renderDefaultChildren from '../../rendering/renderDefaultChildren';
-import { addGTIdentifier, writeChildrenAsObjects } from '../../internal';
-import useGTContext from '../../provider/GTContext';
-import renderTranslatedChildren from '../../rendering/renderTranslatedChildren';
+import renderDefaultChildren from '../rendering/renderDefaultChildren';
+import { addGTIdentifier, writeChildrenAsObjects } from '../internal';
+import useGTContext from '../provider/GTContext';
+import renderTranslatedChildren from '../rendering/renderTranslatedChildren';
 import { useMemo } from 'react';
-import renderVariable from '../../rendering/renderVariable';
+import renderVariable from '../rendering/renderVariable';
 import { hashSource } from 'generaltranslation/id';
-import renderSkeleton from '../../rendering/renderSkeleton';
-import { TranslatedChildren } from '../../types-dir/types';
-import { useable } from '../../promises/dangerouslyUsable';
-import reactUse from '../../utils/use';
+import renderSkeleton from '../rendering/renderSkeleton';
+import { TranslatedChildren } from '../types-dir/types';
+import { useable } from '../promises/dangerouslyUsable';
+import reactUse from '../utils/use';
 
 /**
  * Build-time translation component that renders its children in the user's given locale.
